[new_feature]  rc 客户端bug修改 see http://rb.corp.taobao.com/r/55995 [1] 当重试过程中找不到ns地址跳出时不能返回成功
git-svn-id: http://code.taobao.org/svn/tfs@2042 b722c2b5-6a69-4643-bcd1-1a75825fa998

+++ branches/dev_tair_meta_from_2.2.12/src/new_client/tfs_rc_client_api_impl.h
@@ -94,9 +94,9 @@ namespace tfs
         int64_t save_buf(const char* source_data, const int32_t data_len,
             char* tfs_name_buff, const int32_t buff_len, const char* suffix = NULL);
 
-        int64_t fetch_file(const char* local_file,
+        int fetch_file(const char* local_file,
                        const char* file_name, const char* suffix = NULL);
-        int64_t fetch_buf(int64_t& ret_count, char* buf, const int64_t count,
+        int fetch_buf(int64_t& ret_count, char* buf, const int64_t count,
                      const char* file_name, const char* suffix = NULL);
 
         bool is_hit_local_cache(const char* tfs_name);
